MLEng Replace docker with pipenv running locally
Docker was not performing well and it was too much to give up my IDE.
PyCharm has a decent system for editing notebooks and has an easy way to
fall-back to a notebook running in a browser.

My current workflow is to edit and execute in PyCharm until I have a
reason do jump into a graph are data that isn't showing up well in the
preview. At which point I'll jump into the notebook.